requirements.txt
nltk==3.9.1
#pysentimiento==0.7.3
python-dotenv==1.1.0
PyYAML==6.0.2
Requests==2.32.3
